improved LibraryLoader for nested libraries
fixed infinite recursion in case of cyclic library dependencies
made loading nested libraries optional
added a loadNested option in XML, refactored XmlStore
added UI for the option
l10n

+++ ru.taximaxim.codekeeper.ui/src/ru/taximaxim/codekeeper/ui/properties/DependencyProperties.java
@@ -49,6 +49,7 @@ import ru.taximaxim.codekeeper.ui.Log;
 import ru.taximaxim.codekeeper.ui.UIConsts;
 import ru.taximaxim.codekeeper.ui.UIConsts.FILE;
 import ru.taximaxim.codekeeper.ui.UIConsts.PROJ_PREF;
+import ru.taximaxim.codekeeper.ui.dialogs.ExceptionNotifier;
 import ru.taximaxim.codekeeper.ui.libraries.LibraryUtils;
 import ru.taximaxim.codekeeper.ui.localizations.Messages;
 import ru.taximaxim.codekeeper.ui.prefs.AbstractTxtEditingSupport;
@@ -62,6 +63,7 @@ public class DependencyProperties extends PropertyPage {
 
     private DependenciesListEditor editor;
     private Button btnSafeMode;
+    private Button btnLoadNested;
     private IProject proj;
 
     @Override
@@ -78,22 +80,30 @@ public class DependencyProperties extends PropertyPage {
         Composite area = new Composite(parent, SWT.NONE);
         area.setLayout(new GridLayout());
 
-        editor = new DependenciesListEditor(area);
-        editor.setLayoutData(new GridData(GridData.FILL_BOTH));
-
         List<PgLibrary> input;
+        boolean loadNested;
         try {
             input = store.readObjects();
+            loadNested = store.readLoadNestedFlag();
         } catch (IOException e) {
             Log.log(e);
             input = new ArrayList<>();
+            loadNested = false;
         }
+
+        editor = new DependenciesListEditor(area);
+        editor.setLayoutData(new GridData(GridData.FILL_BOTH));
         editor.setInputList(input);
 
         btnSafeMode = new Button(area, SWT.CHECK);
         btnSafeMode.setText(Messages.DependencyProperties_safe_mode);
         btnSafeMode.setToolTipText(Messages.DependencyProperties_safe_mode_desc);
         btnSafeMode.setSelection(prefs.getBoolean(PROJ_PREF.LIB_SAFE_MODE, true));
+
+        btnLoadNested = new Button(area, SWT.CHECK);
+        btnLoadNested.setText(Messages.DependencyProperties_load_nested);
+        btnLoadNested.setToolTipText(Messages.DependencyProperties_load_dependencies);
+        btnLoadNested.setSelection(loadNested);
 
         return area;
     }
@@ -103,7 +113,7 @@ public class DependencyProperties extends PropertyPage {
         try {
             prefs.putBoolean(PROJ_PREF.LIB_SAFE_MODE, btnSafeMode.getSelection());
             prefs.flush();
-            store.writeObjects(editor.getList());
+            store.writeDependencies(editor.getList(), btnLoadNested.getSelection());
             refreshProject();
             setValid(true);
             setErrorMessage(null);
@@ -126,10 +136,18 @@ public class DependencyProperties extends PropertyPage {
 
             @Override
             public void widgetSelected(SelectionEvent e) {
+                MessageBox mbQuestion = new MessageBox(getShell(), SWT.ICON_QUESTION | SWT.YES | SWT.NO);
+                mbQuestion.setText(Messages.DependencyProperties_clear_cache);
+                mbQuestion.setMessage(Messages.DependencyProperties_clear_cache_descr);
+                if (mbQuestion.open() != SWT.YES) {
+                    return;
+                }
+
                 try {
                     FileUtils.deleteRecursive(LibraryUtils.META_PATH);
                 } catch (IOException ex) {
-                    Log.log(Log.LOG_ERROR, "Error while clear cache", ex); //$NON-NLS-1$
+                    ExceptionNotifier.notifyDefault(Messages.DependencyProperties_clear_cache_error, ex);
+                    return;
                 }
             }
         });
